Added tests for home page and loginpage
