+++ client/package.json
@@ -6,6 +6,8 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.4.0",
     "@testing-library/user-event": "^7.2.1",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.4.1",
     "eslint-config-react-app": "^5.1.0",
     "lucid-ui": "^5.5.0",
     "react": "^16.12.0",
@@ -13,7 +15,9 @@
     "react-lines-ellipsis": "^0.14.1",
     "react-player": "^1.14.2",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.3.0"
+    "react-scripts": "3.3.0",
+    "react-test-renderer": "^16.12.0",
+    "sinon": "^9.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
